# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "پروژه سیستم وبلاگ به صورت RestApi",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/categories/list": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "لیست دسته‌بندی‌ها (خصوصی)",
                "operationId": "008f2f4df73a2917a445b739dc488874",
                "responses": {
                    "200": {
                        "description": "لیست دسته‌بندی‌ها"
                    },
                    "401": {
                        "description": "نیازمند احراز هویت"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "نمایش یک دسته‌بندی (عمومی)",
                "operationId": "d809927cce155b8469de92d4c76f9288",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "دسته‌بندی پیدا شد"
                    },
                    "404": {
                        "description": "دسته‌بندی پیدا نشد"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "ویرایش دسته‌بندی (نیازمند احراز هویت)",
                "operationId": "a85a6c8ff85ebcaf43aaecc48fc4f15b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "دسته‌بندی با موفقیت ویرایش شد"
                    },
                    "401": {
                        "description": "نیازمند احراز هویت"
                    },
                    "404": {
                        "description": "دسته‌بندی پیدا نشد"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "حذف دسته‌بندی (نیازمند احراز هویت)",
                "operationId": "1185a4f5af2bcdbeb427e3c127a15f6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "دسته‌بندی با موفقیت حذف شد"
                    },
                    "401": {
                        "description": "نیازمند احراز هویت"
                    },
                    "404": {
                        "description": "دسته‌بندی پیدا نشد"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/categories/new": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "ایجاد دسته‌بندی جدید (نیازمند احراز هویت)",
                "operationId": "3a4cff9e4b83883e50df97b85e0f088e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "دسته‌بندی با موفقیت ایجاد شد"
                    },
                    "401": {
                        "description": "نیازمند احراز هویت"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/categories/trashed/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "لیست دسته‌بندی‌های حذف‌شده (نیازمند احراز هویت)",
                "operationId": "a1d9f277e61396dad5f5ce3259ad950b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "لیست دسته‌بندی‌های حذف‌شده"
                    },
                    "401": {
                        "description": "نیازمند احراز هویت"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/categories/restore/{id}": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "بازیابی دسته‌بندی حذف‌شده (نیازمند احراز هویت)",
                "operationId": "3da3ae243d76e67d3a1e68f8bb5bf6bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "دسته‌بندی با موفقیت بازیابی شد"
                    },
                    "401": {
                        "description": "نیازمند احراز هویت"
                    },
                    "404": {
                        "description": "دسته‌بندی پیدا نشد"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/categories/force-delete/{id}": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "حذف دائمی دسته‌بندی (نیازمند احراز هویت)",
                "operationId": "b215e403ef8e3e0ccff7c685ce5c6bcb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "دسته‌بندی با موفقیت به صورت دائمی حذف شد"
                    },
                    "401": {
                        "description": "نیازمند احراز هویت"
                    },
                    "404": {
                        "description": "دسته‌بندی پیدا نشد"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}